requirements.txt
numpy
pandas
requests
Pillow
scikit-image
torch
torchvision
